Add hooks for the context() process
This adds hooks into the initialization and teardown of context objects.
You can add hooks via the hub that effect all contexts created for the
hub. You can also add per-context hooks at creation time that will not
apply to future contexts.

Profling shows me that this does not significantly impact performance
when there are no hooks (default).

This also introduces the $ctx->release method, and requires it be called
at the end of a sub that aquires a context. This is done for a couple
reasons:

 * It seems less magical, people will understand it more than garbage
   collection magic (which still happens under the hood) It makes it
   clear that there is an end of life.
 * It provides a point for the hooks to fire outside of a destructor.
   Some hook logic, such as throwin an exception, simply will not work
   in a destructor.
 * Requiring a call to ->release is easy to add now, but hard to add
   later. Having it in place gives us other implementation possiblities
   in the future if we need them.

This also introduces some additional protections, specifically it will
check to make sure you did not persist your context after the scope has
returned, it does this by noting the stack depth where it was created.

Initially I toyed with some XS to get the stack depth, but it turns out
the stack depth check is fast enough in the majority case that the XS
version is actually slower.