Fix IPC race condition
If you fork/thread before running any tests the root hub would not be
initialized, this resulted in each process creating a root hub.

This change makes it so that if both context and IPC are loaded the root
hub gets initialized. This is done in such a way as to make it work
regardless of load order.